+++ l1-contracts/test/base/RollupBase.sol
@@ -72,7 +72,7 @@ contract RollupBase is DecoderBase {
       endFull.block.blockHash,
       bytes32(0), // WHAT ?
       bytes32(0), // WHAT ?
-      bytes32(bytes20(_prover))
+      bytes32(uint256(uint160(bytes20(_prover)))) // Need the address to be left padded within the bytes32
     ];
 
     bytes32[] memory fees = new bytes32[](Constants.AZTEC_MAX_EPOCH_DURATION * 2);
