+++ l1-contracts/test/fees/FeeRollup.t.sol
@@ -452,7 +452,7 @@ contract FeeRollupTest is FeeModelTestPoints, DecoderBase {
 
           burnSum += manaUsed * point.outputs.mana_base_fee_components_in_fee_asset.congestion_cost;
 
-          fees[feeIndex * 2] = bytes32(bytes20(coinbase));
+          fees[feeIndex * 2] = bytes32(uint256(uint160(bytes20(coinbase))));
           fees[feeIndex * 2 + 1] = bytes32(fee);
         }
